Fix debugger background color
When using Firefox in dark mode, `about:blank` has a dark background
instead of white. The debugger popup window uses `about:blank`, and
as such we get the dark background in the debugger. The text is hard to
read then.

When a web page does not set the page background color, it is up to the
user agent to decide the background color. Firefox lets you configure
this background color. So it is a good idea to set the page background
color anyway.

Ideally, we’d have a dark mode for the debugger too, but that is out of
scope for this commit.