Handle large collections in Debugger
Large collections, like lists with thousands of items, either cause the
debugger and the whole application to be very slow, or crash with a
stack overflow.

Closes https://github.com/elm/compiler/issues/2133
Closes https://github.com/elm/browser/issues/90
Closes https://github.com/elm/browser/issues/104
Closes https://github.com/elm/browser/issues/132

Closes https://github.com/elm/browser/pull/120. That’s a previous PR
that touched on this problem. That PR makes the experience go from
crashing to awfully slow.

This PR is fast even with big collections, by expanding just the first
100 items. There’s a “view more” button to show the next 100. I went
with the simplest possible UX around that, since it’s unclear if anyone
actually tries to click to item number 9000 – at that point, search is
better. And then you could just as well do the search with Elm code and
`Debug.log` rather than using some obscure search syntax in the
debugger. I think looking at the model in the debugger is more for
learning the overall structure of an app, than to look at every single
item of long lists, sets or dicts.